REFACTOR

+++ Lab-Reserve/src/main/java/com/edu/nefu/labreserve/repository/PeriodRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PeriodRepository extends JpaRepository<Period, Long> {
+public interface PeriodRepository extends JpaRepository<Period, Integer> {
 
 }
